refactor: Number 클래스 추가
- 음수 validate Number 클래스로 이동
- String to int Number 클래스로 이동
